\name{simMSM-package}
\alias{simMSM-package}
\docType{package}
\title{simMSM}
\description{Provides functions for simulation of event histories with
  non-linear baseline hazard rate functions and non-linear covariate
  effects using inversion sampling on all cause hazard rate functions.}
\details{The central function of this package is
  \code{simeventhistories} which wraps all other functions to perform
  simulation of event histories.}
\author{Holger Reulen}
\references{Beyersmann, J., Schumacher, M., Allignol, A. (2012): 'Competing Risks and
  Multistate Models with R', Springer.

  Kneib, T., Hennerfeind, A. (2008): 'Bayesian semi parametric
  multi-state models', Statistical Modelling, 8, 2, 169--198.}
\keyword{ package }