% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simmsm.R
\name{simmsm}
\alias{simmsm}
\title{Function for Simulating Data for Marginal Structural Model}
\usage{
simmsm(subjects = 2500, tpoints = 10, psi = 0.3, n = 5)
}
\arguments{
\item{subjects}{Number of Subjects in each simulated dataset}

\item{tpoints}{Maximum number of time-points each subjects are followed}

\item{psi}{Causal effect parameter for Marginal Structural Model}

\item{n}{Number of simulated datasets an user wants to generate}
}
\description{
Simulates data suitable for fitting Marginal Structural Model.
}
\author{
Ehsan Karim (R porting from SAS)
}
\references{
Young J.G., Hernan M.A., Picciotto S., and Robins J.M. Relation between three classes of structural models for the effect of a time-varying exposure on survival. Lifetime Data Analysis, 16(1):71-84, 2010.
}
\keyword{Marginal}
\keyword{Model}
\keyword{Simulation,}
\keyword{Structural}

